% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.add_dates_and_colors}
\alias{.add_dates_and_colors}
\title{Assign colors to special dates}
\usage{
.add_dates_and_colors(calendar, special_dates, color_values, fill_values,
  today, week_start)
}
\arguments{
\item{calendar}{A basic life calendar object; output tibble from
\code{.make_basic_calendar}.}

\item{special_dates}{A named list of special dates or intervals.}

\item{color_values}{A named list of colors (to be used for "color" aesthetic
in calendar graphic). The names should correspond to the names used for
\code{special_dates}.}

\item{fill_values}{A named list of colors (to be used for "fill" aesthetic
in calendar graphic). The names should correspond to the names used for
 \code{special_dates}.}

\item{today}{A date object specifying when "today" is. This is used later
to color the "past" and "future" differently, if desired.}

\item{week_start}{Integer; what day is taken as the first day of the week?
(1 = Monday, 7 = Sunday, etc.)}
}
\value{
The given life calendar tibble, with additional columns:
\itemize{
 \item{"outline"}{A character column containing either "default", or else
  the name of the corresponding special date.}
 \item{"fill"}{A character column containing either "default", or "past", or
  the name of the corresponding special date.}
 \item{"outline_thickness"}{A character column containing either "default" or
  "grid" (if the year or week index is a multiple of five; this is for
 optionally including a grid pattern in the calendar graphic.)}
}
Note that each week can have only a single value in each of the color
aesthetics columns. If there are overlapping special dates (e.g. an interval
that overlaps with another, or a date that falls within an interval), the
following rules of priority are followed: date objects always take priority
over interval objects, and dates (or intervals) that are specified later
in the list of special dates take priority over dates (or intervals)
listed sooner.

A week is considered "in" an interval if there is any overlap between the
week and the interval.
}
\description{
Given a basic life calendar and a list of special dates with corresponding
colors, add fields to the calendar for later use with color aesthetics.
Note: a specified date will only be added to the calendar if a corresponding
color is specified in either \code{color_values} or \code{fill_values}.
}
\keyword{internal}
