% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.generate_graphic}
\alias{.generate_graphic}
\title{Make a graphical representation of a life calendar}
\usage{
.generate_graphic(calendar, fill_values, color_values, fill_default,
  fill_past, color_default, default_thickness, grid_thickness, caption,
  ellipse_n)
}
\arguments{
\item{calendar}{A basic calendar object; output tibble from
\code{.add_dates_and_colors}.}

\item{fill_values}{A named list of colors (to be used for "fill" aesthetic
in calendar graphic). The names should correspond to the names used for
 \code{special_dates}.}

\item{color_values}{A named list of colors (to be used for "color" aesthetic
in calendar graphic). The names should correspond to the names used for
\code{special_dates}.}

\item{fill_default}{Character; color to be used for default week fill.}

\item{fill_past}{Character; color to be used to fill weeks in the "past".}

\item{color_default}{Character; color to be used for default outline color.}

\item{default_thickness}{Numeric; default thickness used to draw week cells.}

\item{grid_thickness}{Numeric; thickness used to draw "grid" week cells.}

\item{caption}{Character; caption to included at bottom of graphic.}

\item{ellipse_n}{Integer; number of polygon vertices to use for approximating
ellipses.}
}
\value{
A ggplot object graphically representing the given life calendar.
 Weeks are represented by ellipses, and each year is a single row of
 ellipses (usually 52, though occasionally 53).
}
\description{
Given a life calendar tibble and a list of special dates with corresponding
colors, add fields to the calendar for later use with color aesthetics.
Note: a specified date will only be added to the calendar if a corresponding
color is specified in either \code{color_values} or \code{fill_values}.
}
\keyword{internal}
