% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.make_basic_calendar}
\alias{.make_basic_calendar}
\title{Make a basic life calendar}
\usage{
.make_basic_calendar(birth_date, week_start, num_years)
}
\arguments{
\item{birth_date}{A date object specifying when to begin the life calendar.}

\item{week_start}{Integer; what day is taken as the first day of the week?
(1 = Monday, 7 = Sunday, etc.)}

\item{num_years}{Integer; the number of years included in the calendar.}
}
\value{
A tibble with a row for each week starting at \code{birth_date} and
continuing for \code{num_years} years, with columns:
\itemize{
 \item{"date"}{The start date of the week}
 \item{"year_count"}{The year the week falls in, starting with 1}
 \item{"week_in_year"}{The index of the week in the year, starting with 1}
}
}
\description{
Construct a tibble of weeks, starting from a given birth date, with columns
giving the week start date, year, and week in year. The convention adopted is
that each week is represented by its first day, where the canonical first day
of the week is chosen by setting the \code{week_start} parameter.
}
\details{
This means, for example, that given a birth date that fell on a Monday and
weeks starting on Sundays, the first date in the calendar will be the day
before the given birth date.
}
\keyword{internal}
