% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_calendar.R
\name{make_calendar}
\alias{make_calendar}
\title{Make a life calendar graphic}
\usage{
make_calendar(birth_date, week_start = 7L, num_years = 100L,
  special_dates = NULL, color_values = NULL, fill_values = NULL,
  fill_default = "white", fill_past = "gray60",
  color_default = "gray40", default_thickness = 0.6,
  grid_thickness = 0.85, today = lubridate::today(), caption = NULL,
  ellipse_n = 16L)
}
\arguments{
\item{birth_date}{A date object specifying when to begin the life calendar.}

\item{week_start}{Integer; what day is taken as the first day of the week?
(1 = Monday, 7 = Sunday, etc.)}

\item{num_years}{Integer; the number of years included in the calendar.}

\item{special_dates}{A named list of special dates or intervals.}

\item{color_values}{A named list of colors (to be used for "color" aesthetic
in calendar graphic). The names should correspond to the names used for
\code{special_dates}.}

\item{fill_values}{A named list of colors (to be used for "fill" aesthetic
in calendar graphic). The names should correspond to the names used for
 \code{special_dates}.}

\item{fill_default}{Character; color to be used for default week fill.}

\item{fill_past}{Character; color to be used to fill weeks in the "past".}

\item{color_default}{Character; color to be used for default outline color.}

\item{default_thickness}{Numeric; default thickness used to draw week cells.}

\item{grid_thickness}{Numeric; thickness used to draw "grid" week cells.}

\item{today}{A date object specifying when "today" is. This is used later
to color the "past" and "future" differently, if desired.}

\item{caption}{Character; caption to included at bottom of graphic.}

\item{ellipse_n}{Integer; number of polygon vertices to use for approximating
ellipses.}
}
\value{
A ggplot object graphically representing a life calendar. Weeks are
  represented by ellipses, and each year is a single row of ellipses (usually
  52, though occasionally 53).
}
\description{
This routine produces a life calendar as a ggplot object. A life calendar is
a week-by-week representation of an entire (expect) lifetime. At reasonable
scales, one hundred years' worth of weeks can fit on a single sheet of paper.
The graphic produced by this function is suitable for printing at that scale.
}
\details{
Each week is represented by a small "bubble" (think standardized test answer
sheet). The first week is the week that the subject was born, and a new row
begins with each new birthday week.

At minimum, a life calendar is defined by a birth date (the only function
parameter without a default value). By default, the bubbles are filled in up
to (and including) the current week. This behavior can be controlled by
setting the \code{fill_past} parameter. "Future" bubbles are left unfilled by
default, but this can be overridden by passing a color to the
\code{fill_default} parameter.

Optionally, significant past dates/periods can be colored differently. This
is accomplished by passing a named list of special dates and/or inteverals to
the \code{special_dates} parameter, along with corresponding named lists of
colors to the \code{color_values} and/or \code{fill_values} parameters (to
specify bubble outline and bubble fill colors, respectively).

It is sometimes desirable to add "grid lines" to the array of bubbles by
drawing certain bubble outlines more thickly. This is currently the default;
to *not* show the grid, set the parameters \code{default_thickness} and
\code{grid_thickness} to the same value (suggested 0.6).
}
